+++ packages/pg-gateway/src/index.ts
@@ -1,4 +1,4 @@
 export { default as PostgresConnection } from './connection.js';
 export * from './connection.js';
-export * from './util.js';
 export * from './auth/sasl/scram-sha-256.js';
+export * from './auth/md5.js';
